Improve query API (#15)
* - test

* Delete LICENSE

For Hendrik ;-)

* - renamed alica_test to alica_tests
- fixed task_assignment test although it does not really do anything
- fixed some output errors of the alica logger (now ID instead of ID's address is written into log file)

* style update Planbase, add acessor to find out if planbase is waiting for steppong signal

* factor out unit test main

* fix wrong compare

* Fix authority system and unit test

* avoid unnecessary dereferencing of entypoints

* address review comments

* Fix local agent wrongly indicating a team updat

* log robotids, not their adress

* remove unneeded accessor

* add some constness to agent

* improve robotID debug output

* Do not clear success marks of the local agent all the time.

* prevent doubling of success marks...

* minor improvements

* change tests to rostests, dependency is there anyway.

* address review comments

* add rostest dependency

* remove long time limit again

* Simplify the query interface

* Fix test_alica_problem_composition

* fix task_assignment test

* remove debug flag

* fix wrong filter in synctalk

* Fix to fast path event + test for success from behavior

* fix copy&paste issue

* clean up

* formatting changes

* first simplification iteration on RunningPlan

* initial check-in, WIP

* added .clang_format file

* Update test_alica_authority.cpp

That is more obvious.

* - removed merge artifacts

* copied stuff from old alica_test to alica_tests

* Compiler error fixes, simplify ignoring agents (#8)

* Compiler error fixes, simplify ignoring agents

* remove vscode and add to gitignore

* added clang format file

* moved coding guidelines into style folder

* iteration 2

* Update CodingGuidelines.md

Merging the carpe-noctem-cassel coding guide with the one here in alica

* Update CodingGuidelines.md

update version

* Update CodingGuidelines.md

moved general things upwards + changes namespace examples

* Update CodingGuidelines.md

* Update CodingGuidelines.md

* Update CodingGuidelines.md

* - tryed to fix wrong naming for alica_test(s)
- removed old test package

* - fixed nameing

* Update CodingGuidelines.md

- moved pragma once into header section
- added remark to relative includes

* Update CodingGuidelines.md

* more refactoring

* model constification -> engine compiles

* style update

* remove accidental log file

* clean up

* better whitespaces

* fix bad merge

* format update

* update to compile supplementary and tests

* fix remaining compiler issues in tests

* fix comment issue

* fixing missing return

* more errors mkay

* test improvement after fixing subvariable bug

* fix planwriter issues

* clean planningProblem and destinationPath

* clean capabilityDefinitionSet

* remove unused RoleUsage

* cleaning

* beautification and removal of unnecessary new

* fix missing sort predicate, add test

* cleaning & removing some memory allocations

* reorder header

* wip

* fix extra endif

* beautification

* caps ftw

* consistency

* review comments

* wip

* use proper comparison in sorting

* Initial integration of Variant

* fix typo

* address review comments

* some fixes

* remove unnecessary setter

* - Refactored the way solvers are addressed
- Made Query work with the new solver interface
- Clean AlicaEngine.h somewhat

* simplifications

* add blackboard and hash function

* fixes & use new interface in tests

* raise warning level

* test compile

* fix test

* minor fix

* Variant tests

* add blackboard test

* test for variable syncing, clean up and additional static assert

* const override

* improve BBIdent comparison operators

* review comments

* Update SuccessMarks.h

makes no sense

* refactoring query algorithm, WIP checkin

* cleaning

* align clang format

* rename Sets to Grps

* update types.h

* fix rename

* fix issues with clang format file

* WIP

* Fix compiler errors

* rename

* fix range accessors

* fix unit tests

* fix renaming

* fix bad merge

* clean assignment & fix bug in forallagents

* Query API should only use const RunningPlan

* remove unused functions

* generalize getVariableByName

* fix missing sovlervars incase no problempart is involved

* fix wrong id check

* bug fixes

* add variable handling test

* remove warnings

* make sure agents know each other in the test first

* fix condition

* fix crash bug in synchronization

* give verbose info

* increase test timeout

* minor fixes and give more time to behavior trigger test

* address PR comments

* quick fix

+++ alica_engine/src/engine/PlanRepository.cpp
@@ -17,8 +17,7 @@ PlanRepository::PlanRepository() {}
 PlanRepository::~PlanRepository() {}
 
 bool PlanRepository::verifyPlanBase() const {
-
-    //Every entrypoint has a task:
+    // Every entrypoint has a task:
     for (const EntryPoint* ep : getEntryPoints()) {
         if (ep->getTask() == nullptr) {
             std::cerr << "EntryPoint " << ep->toString() << " does not have a task." << std::endl;
@@ -26,9 +25,9 @@ bool PlanRepository::verifyPlanBase() const {
             return false;
         }
     }
-    //Every plans entryPoints are sorted:
+    // Every plans entryPoints are sorted:
     for (const Plan* p : getPlans()) {
-        for (int i = 0; i < p->getEntryPoints().size() - 1; ++i) {
+        for (int i = 0; i < static_cast<int>(p->getEntryPoints().size()) - 1; ++i) {
             if (p->getEntryPoints()[i]->getId() >= p->getEntryPoints()[i + 1]->getId()) {
                 std::cerr << "Wrong sorting of entrypoints in plan " << p->toString() << std::endl;
                 assert(false);
